api-main/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.2.0
autobahn==24.4.2
Automat==22.10.0
blinker==1.9.0
cachetools==5.5.0
Cerberus==1.3.5
certifi==2024.7.4
cffi==1.17.0
chapa-cli==0.0.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==43.0.0
cssselect2==0.7.0
daphne==2.5.0
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.2.23
django-allauth==0.50.0
django-cors-headers==3.10.1
django-crontab==0.7.1
django-environ==0.9.0
django-filter==22.1
django-model-utils==4.5.1
django-next-gen==0.1.0
django-notifications-hq==1.7.0
django-rest-auth==0.9.5
django-sequences==2.7
django-watson==1.6.3
django_chapa==1.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
Flask==3.1.1
git-filter-repo==2.47.0
google-auth==2.34.0
html5lib==1.1
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.7.0
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonfield==3.1.0
kolo==2.8.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
num2words==0.5.14
oauthlib==3.2.2
oscrypto==1.3.0
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.10
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.19.1
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pyngrok==7.2.9
pyOpenSSL==24.2.1
PyPDF3==1.0.6
pypng==0.20220715.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-decouple==3.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9
service-identity==24.1.0
sib-api-v3-sdk==7.6.0
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
swapper==1.3.0
tinycss2==1.3.0
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.5
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzlocal==5.2
ulid-py==1.1.0
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==5.3.0
xhtml2pdf==0.2.8
zipp==3.22.0
zope.interface==7.0.1
